Move evaluation logic to utils file
